index.html
In the above code, you will need to replace CONTRACT_ADDRESS_GOES_HERE with the actual address of your deployed PlaySudoku smart contract.

This frontend allows the user to connect their Metamask wallet, view the current status of the game board, and verify their win status. The game board is displayed as a 2D grid, and the user can click the "Verify Win" button to check if they have successfully completed the Sudoku puzzle. The win status is displayed below the game board, and will change depending on whether the user has won or not.